requirements.txt
langchain
langchain-anthropic
langchain-community
langchain-core
python-dotenv
pypdf
docx
python-docx
streamlit
anthropic
langchain-openai
pypdf2

